+++ gradio/components/textbox.py
@@ -2,7 +2,6 @@
 
 from __future__ import annotations
 
-import warnings
 from typing import Callable, Literal
 
 import numpy as np
@@ -14,6 +13,7 @@ from gradio.components.base import (
     IOComponent,
     _Keywords,
 )
+from gradio.deprecation import warn_style_method_deprecation
 from gradio.events import (
     Blurrable,
     Changeable,
@@ -124,7 +124,6 @@ class Textbox(
             **kwargs,
         )
         TokenInterpretable.__init__(self)
-        self.cleared_value = ""
         self.type = type
 
     def get_config(self):
@@ -256,9 +255,7 @@ class Textbox(
         """
         This method is deprecated. Please set these arguments in the constructor instead.
         """
-        warnings.warn(
-            "The `style` method is deprecated. Please set these arguments in the constructor instead."
-        )
+        warn_style_method_deprecation()
         if show_copy_button is not None:
             self.show_copy_button = show_copy_button
         if container is not None: